requirements.txt
pyautogui
pytesseract
opencv-python
pillow
keyboard
requests
